backend/requirements.txt
Flask
Flask-Login
Flask-Session
Flask-CORS
Flask-SQLAlchemy
SQLAlchemy
werkzeug
pdfplumber
pypdf
faiss-cpu
requests
python-dotenv
google-generativeai
PyMuPDF